Fix incorrect step output variable in workflow
The step output variable for the site directory for GH Pages is supposed
to be named `site-dir`, but Copilot autocomplete with `site_dir` and
this was missed. This resulted in the whole repo being deployed as a
site page instead of just the `site` directory. Oof. This fixes the
issue.