Separate `build-docs` into new reusable workflow
The `fhenix` project has already experimented a bit with leveraging
a reusable workflow for building the docs pages, which allows for CI
to check that the docs build correctly without actually performing a
deployment. This restructures the workflows in the template so that it
can be more easily instantiated by other repos.